Delay use of Redis to after startup
The Statistics module of the Fleet Manager needs to load data from
Redis. Previously, it was relying on a connection started during
the import of some Python modules, but that was changed in BP-944.

This commit changes the code to only load data from Redis when the
API is first called, rather than during Python import, so that we
can be sure the connection to Redis is already established.